refactor: relocate agent_task_lifecycle_event from runner into core
Prep for the runner extraction. agent_task_lifecycle_event was filed under
runner/ but is pure core logic: it parses api_jobs job events into agent-task
lifecycle events and depends only on core modules (lab_contract, api_jobs,
agent_tasks) — no runner internals.

Moved it to agent_task_lifecycle/ where it belongs. Re-exported from runner::
so runner-internal `super::agent_task_lifecycle_event` /
`crate::runner::agent_task_lifecycle_event` call sites resolve unchanged; core
callers (lifecycle_ops, observation/evidence_report) now reference it via
crate::agent_task_lifecycle::.

This severs 5 core -> runner references (the AgentTaskRunPlanLifecycleEvent type
+ its two builder fns) that would otherwise be core <-> runner cycles once
runner becomes its own crate.

Verified: cargo check clean for -p homeboy-core --lib/--tests, -p homeboy-cli
--lib.
